requirements.txt
alembic==1.15.2
altair==5.5.0
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
category_encoders==2.8.1
certifi==2023.7.22
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cookiecutter==2.6.0
cramjam==2.9.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
email-validator==2.1.0.post1
entrypoints==0.4
executing==2.1.0
Faker==37.1.0
fastapi==0.104.0
fastjsonschema==2.21.1
fastparquet==2024.11.0
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-play-scraper==1.2.7
greenlet==3.1.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
imbalanced-learn==0.13.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
llvmlite==0.44.0
lxml==5.3.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.0
numpy==2.1.2
oauthlib==3.2.2
optuna==4.2.1
orjson==3.9.9
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==10.4.0
platformdirs==4.3.3
plotly==6.0.1
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.47
protobuf==5.29.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
python-multipart==0.0.6
python-slugify==8.0.4
pytrends==4.9.2
pytz==2024.2
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.47.1
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.4
streamlit==1.44.1
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
tweepy==4.14.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
tzdata==2024.1
ujson==5.8.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.23.2
watchdog==6.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
xgboost==3.0.0